tests/pyproject.toml
[tool.poetry]
name = "skand.platform.backend.serverless.image-processing"
version = "1.12.0"
description = "Serverless function used to handle uploading and processing of images"
authors = ["Kevin Lu <kevin@skand.io"]

[tool.poetry.dependencies]
python = "~3.8"
Pillow = "9.1.0"
boto3 = "^1.12.49"
tqdm = "^4.46.0"
piexif = "^1.1.3"
pyavm = "^0.9.4"
xmltodict = "^0.12.0"
pytest = "^6.2.4"
pytest-watch = "^4.2.0"
awscli = "^1.20.33"
moto = "^5.0.20"
pytest-testmon = "^2.0.12"
Jinja2 = "^3.1.2"
weasyprint = "^59.0"
pys3thon = "^0.4.1"
dapr = "1.13.0"
dapr-ext-grpc = "1.12.0"
cloudevents = "^1.11.0"
opentelemetry-sdk = "1.25.0"
opentelemetry-api = "1.25.0"
opentelemetry-distro = "0.46b0"
opentelemetry-exporter-otlp = "1.25.0"
opentelemetry-exporter-otlp-proto-common = "1.25.0"
opentelemetry-exporter-otlp-proto-grpc = "1.25.0"
opentelemetry-exporter-otlp-proto-http = "1.25.0"
opentelemetry-instrumentation = "0.46b0"
opentelemetry-instrumentation-aiohttp-client = "0.46b0"
opentelemetry-instrumentation-asgi = "0.46b0"
opentelemetry-instrumentation-asyncio = "0.46b0"
opentelemetry-instrumentation-botocore = "0.46b0"
opentelemetry-instrumentation-logging = "0.46b0"
opentelemetry-instrumentation-requests = "0.46b0"
opentelemetry-instrumentation-system-metrics = "0.46b0"
opentelemetry-instrumentation-threading = "0.46b0"
opentelemetry-instrumentation-urllib = "0.46b0"
opentelemetry-instrumentation-urllib3 = "0.46b0"
opentelemetry-proto = "1.25.0"
opentelemetry-semantic-conventions = "0.46b0"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
flake8 = "^3.7.9"
jupyter = "^1.0.0"
jupyterlab = "^2.1.2"


[tool.black]
line-length = 92
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[build-system]
requires = ["poetry>=1.0.3"]
build-backend = "poetry.masonry.api"

